@@ -9,19 +9,16 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130618163222) do
+ActiveRecord::Schema.define(version: 20131118103709) do
 
-  create_table "menus", :force => true do |t|
-    t.string   "video_id",   :null => false
-    t.string   "title",      :null => false
-    t.integer  "state",      :null => false
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+  create_table "histories", force: true do |t|
+    t.string   "video_id"
+    t.string   "title"
+    t.integer  "state"
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
-
-  add_index "menus", ["created_at", "state"], :name => "index_menus_on_created_at_and_state"
-  add_index "menus", ["video_id"], :name => "index_menus_on_video_id", :unique => true
 
 end
